MicroServices and MicroFrontends
- Send command from UI
- Receive command in Pubs Service
- Send Command  to Production via Azure ServiceBus
- Handle Command, generate DomainEvent

+++ NotEveryoneDrinksBlonde/Api/BrewUp.Production/BrewUp.Production.Module.Extensions/CustomTypes/BeerQuantity.cs
@@ -1,3 +1,3 @@
 ﻿namespace BrewUp.Production.Module.Extensions.CustomTypes;
 
-public record BeerQuantity(double Value);
+public record BeerQuantity(int Value);
